added all the suggestions
