+++ src/webserver/webserver.h
@@ -14,7 +14,6 @@ class Webserver{
         void createFlashWriterEndpoints();
         void createSequenceEndpoints();
 };
-void readSensors(sensorReadings& readings);
 template<typename Func>
 void sendResponse(AsyncWebServerRequest *request, Func func);
